refactor(core): extract batch machinery into core/batch.js
state.js had absorbed the batch depth/queue/flush-wave machinery and
its maintainability index dropped below the repo's complexity budget
(MI 48.3 < 50). The machinery now lives in core/batch.js — which never
imports state.js (no cycle): state.js calls enqueueIfBatching() from
its scheduler and shares MAX_FLUSH_ITERATIONS from there. Behavior,
public API, tests, and coverage are unchanged; both files are back
within budget and each owns one concern.
